Include the field modulus when calculating the hash value.
This makes the test suite pass again. The problem was that some tests
used a different modulus than other tests. Without the modulus as part
of the hash value, shamir.recombine would mix up the cached
recombination vectors and produce incorrect results.

Also removed the class from the key used for lookup in the
recombination vector cache since it is always included in the hash
value.