Reworked runtime to allow the use of several different fields.
The GF (short for Galois Field) function is called with a modulus as
argument and returns a class. Objects instances from this class will
model elements from the field.

The explicit use of IntegerFieldElement was removed from the runtime
since there can now be several fields in play at any given time. This
has lead to an extra argument to some functions, notably greater_than
which now takes two deferreds as well as a field. The plan is to
remove these arguments again by having the deferreds carry a type in
the form of a reference to a field.