Implemented automatic program counter increment.
The decorator increment_pc is to be used on all method which need a
new program counter. Runtime methods like add which does no
communication need not use this decorator.

When methods in the runtime need to add callbacks, they must use the
callback method on Runtime instead of the usual addCallback method.
The new method ensures that the callback is called with the correct
program counter. Simple callbacks that does not depend on the program
counter may continue to use addCallback. Currently, only methods on
Runtime may be scheduled using the callback method.