Added decorator for wrapper functions.
This fixes the many false warnings emitted by epydoc. The problem was
that epydoc saw the argument list of the wrapper function, which was
typically specified as

  def wrapper(*args, **kwargs)

Epydoc would therefore warn about unknown parameters. Using the
wrapper decorator in viff.util solves this problem and allows correct
documentation to be generated without affecting the code in the normal
case since the wrapper decorator is called at "compile time".