Make run.py images needed by trace2html.py.
It turns out that the coverage file generated by trace2html.py
references an image called "blank.png". But this image is not included
in the output, and this leads to 404 errors on the webserver.

I opted for simply storing the PNG image in run.py since it is not
something we need to change often (probably never) and I felt that it
would be nice to have the data where it is used.