Verify that multiplications really are local.
The tests rely on the fact that an already called Deferred will
execute a callback immediately when it is added with addCallback. The
API does not guarantee this behavior, but the current implementation
works that way. I don't know if this will change if the Deferred class
is implemented in C, as this Twisted ticket suggests:

  http://twistedmatrix.com/trac/ticket/2245