Basic support for pre-processing.
This adds support for pre-processing to the BasicRuntime. The idea is
that a protocol run records the needed data using the preprocess
decorator, and using a description of this data, the runtime can begin
the following executions generating it.

The data needed is recorded based on the methods name and program
counter, so this assumes that every execution produces the same tree
of program counters. This is true for simple programs like the
benchmark, but need not be true for programs that branch on the input
data. Please see Issue 3: http://tracker.viff.dk/issue3.