Fixed memory leak.
It turned out that we kept references to empty deques in incoming_data
without ever releasing them. This caused the memory usage to grow as
more and more communication was done.

Checking for empty deques every time something is received or sent is
expensive -- I measured a performance drop of 25%! To make things
worse, I observed that the deques are mostly empty, so we are probably
deleting and recreating them all the time. Patches with optimizations
are very welcome! :-)