Make players attempt to reconnect if connection fails.
It is only if the initial connection fails that they will reconnect,
if the connection fails in the middle of a protocol run they will
still crash. This is probably the right thing to do since we generally
wouldn't know how to recover from a crashed protocol.

This allows the players to be started in any order and is thus a fix
for Issue 8.