New deep_wait function.
This function (it is really a class which acts like a function) takes
a value as input and gives a Deferred back. The promise is that the
Deferred will only trigger when all Deferreds in the value have been
triggered themselves. This happens recursively, meaning that deep_wait
will keep waiting until the value if fully evaluated.